+++ routes/staffRoute.mjs
@@ -13,9 +13,17 @@ router.route("/").get(async(req,res)=>{
     }
 }).post(async(req,res)=>{
     try{
-        const postStaff = new Staff(req.body)
-        await postStaff.save()
-        res.json(postStaff)
+        if(!req.body.name.trim() || !req.body.jobTitle.trim()){
+            res.send("Plankton, you need a name and job title field like this {name: Plankton, jobTitle: ruler of the universe}")
+        }
+        else if(req.body.name.trim().toLowerCase().includes("plankton")){
+            res.send("We know it's you plankton, Mr.Krab's kidnapping karen as we speak unless you log off.")
+        }
+        else{
+            const postStaff = new Staff(req.body)
+            await postStaff.save()
+            res.json(postStaff)
+        }
     }catch(err){
         console.error(err)
         res.status(500).send('Plankton is trying to access our Server. Unable to post staff data.')
@@ -25,7 +33,11 @@ router.route("/").get(async(req,res)=>{
 router.route("/:id").get(async(req,res)=>{
     try{
         const getStaff = await Staff.findById(req.params.id)
-        res.json(getStaff)
+        if(!getStaff){
+            res.send(`${req.params.id} doesn't exist in this database Plankton. Try Again`)
+        } else{
+            res.json(getStaff)
+        }
     }catch(err){
         console.error(err)
         res.status(500).send('Plankton is trying to access our Server. Unable to get staff data by id.')
@@ -33,16 +45,29 @@ router.route("/:id").get(async(req,res)=>{
     
 }).put(async(req,res)=>{
     try{
-        const putStaff = await Staff.findByIdAndUpdate(req.params.id, req.body, {new: true})
-        res.json(putStaff)
+        if(!req.body.name.trim() || !req.body.jobTitle.trim()){
+            res.send("Plankton put the name and/or jobTitle back. Come on. Don't make me call Mr.Krabs on you.")
+        }
+        else if(req.body.name.trim().toLowerCase().includes("plankton")){
+            res.send("We warned you plankton. Say bye bye to Karen. She's our wife now.")
+        }
+        else{
+            const putStaff = await Staff.findByIdAndUpdate(req.params.id, req.body, {new: true})
+            res.json(putStaff)
+        }
     }catch(err){
         console.error(err)
         res.status(500).send('Plankton is trying to access our Server. Unable to update staff data.')
     }
 }).delete(async(req,res)=>{
     try{
-        await Staff.findByIdAndDelete(req.params.id)
-        res.send(`${req.params.id} has been deleted`)
+        const getStaff = await Staff.findById(req.params.id)
+        if(!getStaff){
+            res.send(`Not you tryna delete something that doesn't exist. Try again Plankton`)
+        } else{
+            await Staff.findByIdAndDelete(req.params.id)
+            res.send(`${req.params.id} has been deleted`)
+        }
     }catch(err){
         console.error(err)
         res.status(500).send('Plankton is trying to access our Server. Unable to delete staff data.')
